Added doubles, made int backed by longs

+++ qilletni-api/src/main/java/is/yarr/qilletni/api/lang/types/IntType.java
@@ -1,7 +1,7 @@
 package is.yarr.qilletni.api.lang.types;
 
 public non-sealed interface IntType extends QilletniType {
-    int getValue();
+    long getValue();
 
-    void setValue(int value);
+    void setValue(long value);
 }
